Hookup filter and tls context to authn plugin. Remove authn policy from httpOpts.
